requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.2,<4.0
psycopg2>=2.8.6,<3.0
drf-spectacular>=0.14.1,<1.0
django-cors-headers>=3.7.0,<4.0
